detect root status and complain, fix cfg cp
