+++ bak/__init__.py
@@ -1,1 +1,1 @@
-BAK_VERSION = "0.2.2a5"
+BAK_VERSION = "0.2.2a6"
