+++ setup.py
@@ -8,7 +8,7 @@ require = ['click==7.1.2',
            'rich==9.1.0']
 
 setup(name='bak',
-      version='0.2.2a5',
+      version='0.2.2a6',
       description='the .bak manager',
       author='ChanceNCounter',
       author_email='ChanceNCounter@icloud.com',
@@ -17,7 +17,7 @@ setup(name='bak',
       install_requires=require,
       entry_points='''
       [console_scripts]
-      bak=bak.__main__:bak''',
+      bak=bak.__main__:run_bak''',
       license='MIT License',
       url='https://github.com/bakfile/bak')
 
@@ -29,10 +29,14 @@ except KeyError:
 
 config_file = os.path.join(config_dir, 'bak.cfg')
 default_config = os.path.join(config_dir, 'bak.cfg.default')
+system_default_config = os.path.join('/etc/xdg', 'bak.cfg.default')
 
 if not os.path.exists(config_file):
     copy2('bak/default.cfg', config_file)
 
 if not os.path.exists(default_config):
     copy2('bak/default.cfg', default_config) 
-
+    try:
+        copy2('bak/default.cfg', system_default_config)
+    except PermissionError:
+        pass
